Moved the cash flow components into their own directory
